backend/requirements.txt
alembic==1.7.6
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
pip==22.0.4
pycparser==2.21
PyJWT==2.3.0
setuptools==60.9.3
SQLAlchemy==1.4.32
Werkzeug==2.0.3
wheel==0.37.1
zipp==3.7.0
